refactor(file-search): consolidate overfragmented manager classes
Consolidated 4 overly fragmented manager classes into 3 cohesive units to eliminate excessive separation of concerns:

- Phase 1: Merged SubdirectoryFilterManager and RootFilterManager into FileFilterManager
- Phase 2: Merged HoverStateManager into CursorHighlightManager
- Phase 3: Inlined SuggestionPositionCalculator logic into SuggestionListManager

This reduces the file-search module from 30 to 26 manager classes while improving code navigation and reducing callback-based dependencies. All type checking, linting, and 497 tests pass.
